Created AutoRoutineChooser
Previously I made a command for this, but I found it would make more sense to create an object that is not limited to the abilities of a command so that it could be referenced in Robot, etc.
Everything from the command was transfered and I made some tweeks to the AutoCommand in RobotContainer.